move server code into preview directory
